requirements.txt
astroid==1.6.5
click==6.7
gitdb2==2.0.3
GitPython==2.1.10
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pylint==1.9.2
six==1.11.0
smmap2==2.0.3
wrapt==1.10.11
